requirements/base.txt
git+git://github.com/liqd/adhocracy4.git@0dc3eae28531963bcd420452d6c7b537eef43dfb#egg=adhocracy4
bcrypt==3.1.6
brotli==1.0.7
django-cachalot==2.1.0
django-capture-tag==1.0
django-cloudflare-push==0.2.0
django-sites==0.10
django_csp==3.5
djangorestframework-filters==0.11.1
raven==6.10.0
requests==2.21.0
wagtail==2.3 # pyup: <2.4
whitenoise==4.1.2
zeep==3.3.1

# Inherited a4-core requirements
bleach==3.1.0
Django==2.2.2 # pyup: <2.3
django-allauth==0.39.1
django-autoslug==1.9.4
django-background-tasks==1.2.0
django-ckeditor==5.7.1
django-filter==2.1.0
django-multiselectfield==0.1.8
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
easy-thumbnails==2.6
html5lib==1.0.1
jsonfield==2.0.2
python-dateutil==2.8.0
python-magic==0.4.15
rules==2.0.1
XlsxWriter==1.1.8
